Realapis

+++ DataAccess/Entities/VpgsEvaluation.cs
@@ -7,7 +7,7 @@ public partial class VpgsEvaluation
 {
     public int VpgsevalId { get; set; }
     public int? EvaluationId { get; set; }
-    public decimal ScientificScore { get; set; }
+    public int ScientificScore { get; set; }
     public int StatusId { get; set; }
 
     public virtual Evaluation Evaluation { get; set; } = null!;
